requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
groq==0.13.0
psutil==6.1.0
pytz==2024.2